requirements.txt

fastapi
uvicorn
sqlalchemy
pydantic[email]
passlib[argon2]
argon2-cffi
python-jose
python-multipart
